Move buildKvCachePoolPointers to thop/attentionOp.h
