+++ cpp/tensorrt_llm/thop/thUtils.h
@@ -95,51 +95,6 @@ std::optional<float> getFloatEnv(char const* name);
 
 cudaDataType_t convert_torch_dtype(torch::ScalarType dtype);
 
-struct KvCachePoolPointers
-{
-    void* primaryPoolPtr{nullptr};
-    void* secondaryPoolPtr{nullptr};
-    void* primaryBlockScalePoolPtr{nullptr};
-    void* secondaryBlockScalePoolPtr{nullptr};
-};
-
-inline KvCachePoolPointers buildKvCachePoolPointers(at::Tensor const& hostKvCachePoolPointers, int32_t poolIndex,
-    int64_t intraPoolOffset, int64_t blockSize, int32_t layerIdxInCachePool, int32_t kvFactor, bool isFp4KvCache)
-{
-    KvCachePoolPointers pointers;
-    if (isFp4KvCache)
-    {
-        // For NVFP4 KV cache, extra block scales are stored in separate pools.
-        // The layout of host_kv_cache_pool_pointers is [num_pools, 2 (primary and secondary), 2 (data and scale)].
-        TORCH_CHECK(hostKvCachePoolPointers.dim() == 3);
-        pointers.primaryPoolPtr = reinterpret_cast<void*>(
-            reinterpret_cast<char*>(hostKvCachePoolPointers.index({poolIndex, 0, 0}).item<int64_t>())
-            + intraPoolOffset);
-        pointers.secondaryPoolPtr = reinterpret_cast<void*>(
-            reinterpret_cast<char*>(hostKvCachePoolPointers.index({poolIndex, 1, 0}).item<int64_t>())
-            + intraPoolOffset);
-        // NVFP4 block scaling uses a fixed vector size of 16.
-        auto constexpr vectorSize = 16;
-        auto const bytesPerBlockSf = blockSize / vectorSize * 1 /*bytes per E4M3 sf*/;
-        auto const intraPoolOffsetSf = layerIdxInCachePool * kvFactor * bytesPerBlockSf;
-        pointers.primaryBlockScalePoolPtr = reinterpret_cast<void*>(
-            reinterpret_cast<char*>(hostKvCachePoolPointers.index({poolIndex, 0, 1}).item<int64_t>())
-            + intraPoolOffsetSf);
-        pointers.secondaryBlockScalePoolPtr = reinterpret_cast<void*>(
-            reinterpret_cast<char*>(hostKvCachePoolPointers.index({poolIndex, 1, 1}).item<int64_t>())
-            + intraPoolOffsetSf);
-    }
-    else
-    {
-        TORCH_CHECK(hostKvCachePoolPointers.dim() == 2);
-        pointers.primaryPoolPtr = reinterpret_cast<void*>(
-            reinterpret_cast<char*>(hostKvCachePoolPointers.index({poolIndex, 0}).item<int64_t>()) + intraPoolOffset);
-        pointers.secondaryPoolPtr = reinterpret_cast<void*>(
-            reinterpret_cast<char*>(hostKvCachePoolPointers.index({poolIndex, 1}).item<int64_t>()) + intraPoolOffset);
-    }
-    return pointers;
-}
-
 } // namespace torch_ext
 
 TRTLLM_NAMESPACE_END
